Split a file of utils and move into utils directory
