+++ src/components/CorrectAnswer.vue
@@ -3,7 +3,7 @@ import KeyList from '@/components/KeyList.vue'
 import ShortcutDescription from '@/components/ShortcutDescription.vue'
 import KeyCombination from '@/models/keyCombination'
 import GameKey from '@/stores/gameKey'
-import { injectStrict } from '@/utils'
+import { injectStrict } from '@/utils/injectStrict'
 
 const { state } = injectStrict(GameKey)
 </script>
