+++ src/components/QuestionShow.vue
@@ -2,7 +2,7 @@
 import { XCircleIcon } from '@heroicons/vue/outline'
 
 import GameKey from '@/stores/gameKey'
-import { injectStrict } from '@/utils'
+import { injectStrict } from '@/utils/injectStrict'
 
 const { state } = injectStrict(GameKey)
 </script>
