+++ src/components/RequestSelfScoring.vue
@@ -2,7 +2,7 @@
 import Key from '@/components/Key.vue'
 import SelfScoring from '@/components/SelfScoring.vue'
 import GameKey from '@/stores/gameKey'
-import { injectStrict } from '@/utils'
+import { injectStrict } from '@/utils/injectStrict'
 
 const { state } = injectStrict(GameKey)
 </script>
