+++ src/utils/toggleFullscreen.ts
@@ -1,0 +1,9 @@
+export default function toggleFullscreen() {
+  if (!document.fullscreenElement) {
+    document.documentElement.requestFullscreen()
+  } else {
+    if (document.exitFullscreen) {
+      document.exitFullscreen()
+    }
+  }
+}
